Move writer tests to appropriate package
